Implement governed no-effect Codex turn (#102)
Delivers the accepted #102 v7 slice to epic/98-codex-tracer with exact-head technical, packaged macOS, accessibility, visual, audit, and review evidence complete. The inactive lifecycle coordinator failure is the known out-of-scope defect #147; current issue and PR contracts are green.

+++ native/crates/keiko-host-macos/src/request_timing.rs
@@ -7,7 +7,7 @@ pub(crate) struct InFlight {
     pub(crate) cancelled_at_ms: Option<u64>,
     pub(crate) generation: u64,
     pub(crate) started_at_ms: u64,
-    pub(crate) timeout_ms: u16,
+    pub(crate) timeout_ms: u32,
 }
 
 #[derive(Debug)]
